+++ src/main/java/com/onetuks/goguma_bookstore/author/controller/dto/request/AuthorEditRequest.java
@@ -1,0 +1,10 @@
+package com.onetuks.goguma_bookstore.author.controller.dto.request;
+
+import com.onetuks.goguma_bookstore.author.service.dto.param.AuthorEditParam;
+
+public record AuthorEditRequest(String nickname, String introduction) {
+
+  public AuthorEditParam to() {
+    return new AuthorEditParam(this.nickname, this.introduction);
+  }
+}
