+++ src/main/java/com/onetuks/goguma_bookstore/author/service/dto/param/AuthorEditParam.java
@@ -1,0 +1,3 @@
+package com.onetuks.goguma_bookstore.author.service.dto.param;
+
+public record AuthorEditParam(String nickname, String introduction) {}
