+++ src/test/java/com/onetuks/goguma_bookstore/author/service/dto/result/AuthorEditResultTest.java
@@ -1,0 +1,40 @@
+package com.onetuks.goguma_bookstore.author.service.dto.result;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+import com.onetuks.goguma_bookstore.IntegrationTest;
+import com.onetuks.goguma_bookstore.auth.model.Member;
+import com.onetuks.goguma_bookstore.auth.repository.MemberRepository;
+import com.onetuks.goguma_bookstore.auth.vo.RoleType;
+import com.onetuks.goguma_bookstore.author.model.Author;
+import com.onetuks.goguma_bookstore.author.repository.AuthorJpaRepository;
+import com.onetuks.goguma_bookstore.fixture.AuthorFixture;
+import com.onetuks.goguma_bookstore.fixture.MemberFixture;
+import java.io.IOException;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+class AuthorEditResultTest extends IntegrationTest {
+
+  @Autowired private MemberRepository memberRepository;
+  @Autowired private AuthorJpaRepository authorJpaRepository;
+
+  @Test
+  @DisplayName("작가 프로필 수정 엔티티에서 결과 객체로 변환한다.")
+  void fromTest() throws IOException {
+    // Given
+    Member member = memberRepository.save(MemberFixture.create(RoleType.AUTHOR));
+    Author author = authorJpaRepository.save(AuthorFixture.create(member));
+
+    // When
+    AuthorEditResult result = AuthorEditResult.from(author);
+
+    // Then
+    assertAll(
+        () -> assertThat(result.profileImgUrl()).isEqualTo(author.getProfileImgUrl()),
+        () -> assertThat(result.nickname()).isEqualTo(author.getNickname()),
+        () -> assertThat(result.introduction()).isEqualTo(author.getIntroduction()));
+  }
+}
